+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Enhancements
 
 ### Bug Fixes
+* Fixed WFileWidget and WMultiFileWidget to accept filetypes of both MIME and extension type #1584
 
 ## Release 1.5.13