+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 
 ### Enhancements
 
+* Implement dependency vulnerability scan.
+* Update dependencies to remove vulnerabilities.
+
 ### Bug Fixes
 * Fixed WFileWidget and WMultiFileWidget to accept filetypes of both MIME and extension type #1584
+* Fix issue rendering custom class names in WTree #1589.
 
 ## Release 1.5.13