+++ CHANGELOG.md
@@ -6,10 +6,17 @@
 
 ### Enhancements
 
+### Bug Fixes
+
+## Release 1.5.14
+
+### Enhancements
+
 * Implement dependency vulnerability scan.
 * Update dependencies to remove vulnerabilities.
 
 ### Bug Fixes
+
 * Fixed WFileWidget and WMultiFileWidget to accept filetypes of both MIME and extension type #1584
 * Fix issue rendering custom class names in WTree #1589.